refactor base modules and start test event validator
